+++ app/src/main/java/com/serhiiromanchuk/utilitybills/di/component/InsertServiceScreenComponent.kt
@@ -17,7 +17,7 @@ interface InsertServiceScreenComponent {
     @Subcomponent.Factory
     interface Factory {
         fun create(
-            @BindsInstance utilityServiceId: Int,
+            @BindsInstance utilityServiceId: Long,
             @BindsInstance address: String,
         ) : InsertServiceScreenComponent
     }
