+++ app/src/main/java/com/serhiiromanchuk/utilitybills/domain/usecase/bill/DeleteBillItemUseCase.kt
@@ -7,7 +7,7 @@ class DeleteBillItemUseCase @Inject constructor(
     private val repository: BillRepository
 ) {
 
-    suspend operator fun invoke(billItemId: Int) {
+    suspend operator fun invoke(billItemId: Long) {
         repository.deleteBillItem(billItemId)
     }
 }
