+++ app/src/main/java/com/serhiiromanchuk/utilitybills/domain/usecase/utility_service/GetUtilityServiceUseCase.kt
@@ -8,7 +8,7 @@ class GetUtilityServiceUseCase @Inject constructor(
     private val repository: UtilityServiceRepository
 ) {
 
-    suspend operator fun invoke(utilityServiceId: Int): UtilityServiceItem {
+    suspend operator fun invoke(utilityServiceId: Long): UtilityServiceItem {
         return repository.getUtilityService(utilityServiceId)
     }
 }
